[sn32] spi: split the master functions
