+++ os/hal/ports/SN32/LLD/SN32F24xB/SPI/SPI_master.h
@@ -1,0 +1,2 @@
+extern void SPI0_Write(unsigned char *p, int len);
+extern void SPI0_Read3(unsigned char b1, unsigned char b2, unsigned char *b3);
